requirements.txt
python-dotenv
langchain
openai
googlesearch-python
nest_asyncio
tiktoken
faiss-cpu
